06-server

+++ client-admin/src/components/TableRowList.js
@@ -4,13 +4,35 @@ import { useDispatch } from "react-redux"
 import { deleteEffectItem, fetchEffectOneItem } from "../store/actions/itemAction"
 import { useState } from "react"
 import FormModalEdit from "./FormModalEdit"
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
 
 const TableRowList = ({item,index}) => {
+  const MySwal = withReactContent(Swal)
   const [modalShow, setModalShow] = useState(false);
   const dispatch = useDispatch()
   const deleteHandle = (e,id) => {
-    e.preventDefault();
-    dispatch(deleteEffectItem(id))
+    MySwal.fire({
+      title: "Are you sure want to delete?",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonColor: "#0000FF",
+      cancelButtonColor: "#d33",
+      cancelButtonText: "Cancel",
+      confirmButtonText: "Delete",
+    }).then((logout)=>{
+      if (logout.isConfirmed) {
+        MySwal.fire({
+          position: "center",
+          icon: "success",
+          title: "Delete Success",
+          showConfirmButton: false,
+          timer: 1500,
+        });
+        e.preventDefault();
+        dispatch(deleteEffectItem(id))
+      }
+    })
   }
 
   const handleClickEdit = (e,id) => {
@@ -29,7 +51,7 @@ const TableRowList = ({item,index}) => {
         <td>
         {
           item.Ingredients?.map((ingredient,index)=>{
-            return <> <p>- {ingredient.name}</p>
+            return <> <p key={index}>- {ingredient.name}</p>
             </>
           })
         }
